+++ src/Calendar/Calendar.test.js
@@ -52,7 +52,7 @@ describe('<Calendar />', () => {
         expect(wrapper.state('showMonths')).toBeFalsy();
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(1)
             .simulate('click');
@@ -61,7 +61,7 @@ describe('<Calendar />', () => {
 
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(1)
             .simulate('click');
@@ -81,7 +81,7 @@ describe('<Calendar />', () => {
         //open month overlay
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(1)
             .simulate('click');
@@ -110,7 +110,7 @@ describe('<Calendar />', () => {
         //open months view
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(1)
             .simulate('click');
@@ -133,7 +133,7 @@ describe('<Calendar />', () => {
         expect(wrapper.state('showYears')).toBeFalsy();
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(2)
             .simulate('click');
@@ -142,7 +142,7 @@ describe('<Calendar />', () => {
 
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(2)
             .simulate('click');
@@ -156,7 +156,7 @@ describe('<Calendar />', () => {
         expect(wrapper.state('showYears')).toBeFalsy();
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(2)
             .simulate('click');
@@ -204,7 +204,7 @@ describe('<Calendar />', () => {
         expect(wrapper.state('showYears')).toBeFalsy();
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(2)
             .simulate('click');
@@ -230,7 +230,7 @@ describe('<Calendar />', () => {
 
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(0)
             .simulate('click');
@@ -241,7 +241,7 @@ describe('<Calendar />', () => {
         // previous button when year shown
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(2)
             .simulate('click');
@@ -250,7 +250,7 @@ describe('<Calendar />', () => {
 
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(0)
             .simulate('click');
@@ -267,7 +267,7 @@ describe('<Calendar />', () => {
 
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(3)
             .simulate('click');
@@ -278,7 +278,7 @@ describe('<Calendar />', () => {
         // previous button when year shown
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(2)
             .simulate('click');
@@ -287,7 +287,7 @@ describe('<Calendar />', () => {
 
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(3)
             .simulate('click');
@@ -305,7 +305,7 @@ describe('<Calendar />', () => {
 
         wrapper
             .find(
-                'header.fd-calendar__header button.fd-button--light.fd-button--compact'
+                'header.fd-calendar__header button.fd-button--transparent.fd-button--compact'
             )
             .at(3)
             .simulate('click');