+++ src/Forms/FormSelect.js
@@ -7,7 +7,6 @@ const FormSelect = React.forwardRef(({ disabled, children, className, compact, d
 
     useEffect(() => {
         if (!disableStyles) {
-            require('fundamental-styles/dist/fonts.css');
             require('fundamental-styles/dist/form-select.css');
         }
     }, []);